% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Export_Report}
\alias{Export_Report}
\title{Exports a report on crosstab in Word.}
\usage{
Export_Report(
  Crosstabs,
  Outputs,
  File_Name,
  Name_Group,
  Template,
  Document_Title,
  Type,
  Demographic_Category,
  API_Key,
  Group1
)
}
\description{
Exports a report on crosstab in Word.
}
