% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Responses_to_Text}
\alias{Responses_to_Text}
\title{Converts numeric responses to text responses.}
\usage{
Responses_to_Text(Responses, ColumnName, Codebook)
}
\description{
Converts numeric responses to text responses.
}
