% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Results}
\alias{Results}
\title{Returns tables and reports in Microsoft Word and Excel.}
\usage{
Results(
  Dataset,
  Exports,
  Group_1 = c(Treatment, Time, Weight),
  Group_2 = c(Treatment, Time, Weight),
  Template,
  API,
  Significance
)
}
\arguments{
\item{Dataset}{is name of the Excel dataset file.}

\item{Exports}{is optional and the location of where files should be printed to.}

\item{Group_1}{is a vector containing the treatment group "Participants", time "T1", and weighting variable "Weight 1".}

\item{Group_2}{is a vector containing the treatment group "Nonparticipants", time "T2", and weighting variable "Weight 2".}

\item{Template}{is the Word document that serves as the base from which documents will be printed.}

\item{API}{is an optional API key to call ChatGPT to improve report phrasing.}

\item{Significant}{is an optional value for alpha above which significant tests will not be reported.}
}
\description{
This function returns three types of outputs.
\enumerate{
\item Tables on all nominal data.
\item Tables on all ordinal data by all nominal data.
\item A report on all ordinal data by select nominal data.
This function required a dataset formatted using the "Dataset" function.
}
}
