% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Test_T}
\alias{Test_T}
\title{Defines a function that runs a two sample t-test and adds significance markers.}
\usage{
Test_T(ValuetoMark, Group1, Group2, Weight1, Weight2, Paired)
}
\description{
Defines a function that runs a two sample t-test and adds significance markers.
}
